+++ Example/channel_server/priv/static/js/app.js
@@ -1,8 +1,0 @@
-import {Socket} from "deps/phoenix/web/static/js/phoenix"
-
-let socket = new Socket("/socket")
-socket.connect()
-//let chan = socket.channel("rooms:lobby", {})
-//chan.join().receive("ok", chan => {
-//  console.log("Welcome to Phoenix Chat!")
-//})