+++ Example/channel_server/test/controllers/page_controller_test.exs
@@ -1,8 +1,0 @@
-defmodule ChannelServer.PageControllerTest do
-  use ChannelServer.ConnCase
-
-  test "GET /" do
-    conn = get conn(), "/"
-    assert html_response(conn, 200) =~ "Welcome to Phoenix!"
-  end
-end