+++ Example/channel_server/test/views/layout_view_test.exs
@@ -1,3 +1,0 @@
-defmodule ChannelServer.LayoutViewTest do
-  use ChannelServer.ConnCase, async: true
-end