+++ Example/channel_server/test/views/page_view_test.exs
@@ -1,3 +1,0 @@
-defmodule ChannelServer.PageViewTest do
-  use ChannelServer.ConnCase, async: true
-end