+++ Example/channel_server/web/controllers/page_controller.ex
@@ -1,7 +1,0 @@
-defmodule ChannelServer.PageController do
-  use ChannelServer.Web, :controller
-
-  def index(conn, _params) do
-    render conn, "index.html"
-  end
-end