+++ Example/channel_server/web/views/layout_view.ex
@@ -1,3 +1,0 @@
-defmodule ChannelServer.LayoutView do
-  use ChannelServer.Web, :view
-end