+++ Example/channel_server/web/views/page_view.ex
@@ -1,3 +1,0 @@
-defmodule ChannelServer.PageView do
-  use ChannelServer.Web, :view
-end