+++ Pod/Classes/PhxSocket.h
@@ -25,10 +25,10 @@
 @property (nonatomic, readwrite) BOOL reconnectOnError;
 
 - (id)initWithURL:(NSURL*)url;
-- (id)initWithURL:(NSURL*)url params:(NSDictionary*)params;
-- (id)initWithURL:(NSURL*)url params:(NSDictionary*)params heartbeatInterval:(int)interval;
+- (id)initWithURL:(NSURL*)url heartbeatInterval:(int)interval;
 
 - (void)connect;
+- (void)connectWithParams:(NSDictionary*)params;
 - (void)disconnect;
 - (void)reconnect;