+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Changelog
 
+## v0.3.0
+* Decoupled init from connect. You must explicitally call connect.
+* connect now takes socket params.
+
+## v0.2.0
+* Updated to Pass Socket Params to initializer
+
 ## v0.1.0
 * Initial Release.